// packages/daemon/src/generated/signalr/Source.Features.RuntimeBootstrap.Contracts.ts
export type BootstrapPayloadV2 = {
    /** Transpiled from string */
    version: string;
    /** Transpiled from Source.Features.RuntimeBootstrap.Contracts.RuntimeSpecV2 */
    runtimeSpec: RuntimeSpecV2;
    /** Transpiled from System.Collections.Generic.List<Source.Features.RuntimeBootstrap.Contracts.EnvVar> */
    envVars: EnvVar[];
    /** Transpiled from Source.Features.RuntimeBootstrap.Contracts.HooksConfig? */
    hooks?: HooksConfig;
    /** Transpiled from System.Collections.Generic.List<Source.Features.RuntimeBootstrap.Contracts.McpServer> */
    mcps: McpServer[];
    /** Transpiled from Source.Features.RuntimeBootstrap.Contracts.RepoConfig? */
    repo?: RepoConfig;
    /** Transpiled from string? */
    modelSlug?: string;
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.EnvVar */
export type EnvVar = {
    /** Transpiled from string */
    key: string;
    /** Transpiled from string */
    value: string;
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.HooksConfig */
export type HooksConfig = {
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.McpServer */
export type McpServer = {
    /** Transpiled from string */
    name: string;
    /** Transpiled from string */
    url: string;
    /** Transpiled from string */
    scope: string;
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.RepoAccessToken */
export type RepoAccessToken = {
    /** Transpiled from string */
    token: string;
    /** Transpiled from System.DateTime */
    expiresAt: (Date | string);
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.RepoConfig */
export type RepoConfig = {
    /** Transpiled from string */
    url: string;
    /** Transpiled from string */
    branch: string;
    /** Transpiled from string? */
    deployKey?: string;
}
// ...
export type RuntimeSpecV2 = {
    /** Transpiled from int */
    version: number;
    /** Transpiled from string? */
    install?: string;
    /** Transpiled from string? */
    installVerify?: string;
    /** Transpiled from System.Collections.Generic.List<Source.Features.RuntimeBootstrap.Contracts.ServiceSpec>? */
    services?: ServiceSpec[];
    /** Transpiled from string? */
    setup?: string;
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.ServiceSpec */
export type ServiceSpec = {
    /** Transpiled from string */
    name: string;
    /** Transpiled from string */
    command: string;
    /** Transpiled from string? */
    user?: string;
    /** Transpiled from bool */
    autorestart?: boolean;
    /** Transpiled from System.Collections.Generic.Dictionary<string, string>? */
    env?: Partial<Record<string, string>>;
    /** Transpiled from Source.Features.RuntimeBootstrap.Contracts.HealthcheckSpec? */
    healthcheck?: HealthcheckSpec;
    /** Transpiled from string? */
    install?: string;
    /** Transpiled from string? */
    installVerify?: string;
}